reorganize python files, and convert to python3
