+++ README.md
@@ -1,0 +1,3 @@
+# Rush Hour Solver 
+
+Solves your rush hour puzzles.
