+++ rushhour/test/test_position.py
@@ -1,0 +1,4 @@
+from rushhour.position import Position
+
+test1 = Position(1, 1) + Position(2, 2) == Position(3, 3)
+print("Positions can be added " + str(test1))
